refactor: keep Dal models together
