+++ deployer/src/lib.rs
@@ -5,8 +5,7 @@ use std::time::Duration;
 use async_trait::async_trait;
 use bollard::{Docker, API_DEFAULT_VERSION};
 use chrono::Utc;
-use dal::{Dal, Service};
-use deployment::Deployment;
+use dal::{Dal, Deployment, Service};
 use derive_builder::Builder;
 use error::{Error, Result};
 use futures::TryFutureExt;
